rename config scoped var to sensorConfig
